Library Virtual Assistant Borrowing Bot
This borrowing bot was created as part of a project for the NCSS Summer School. It uses a storage API to get and store user data as a JSON file, and also uses Flask to function as a server. This bot, along with the other bots that make up the library assistant, can be interacted with at chat.ncss.cloud/syd-group3

Commands that can be used for the bot:
- borrow {book Title}: borrows a book and adds it to your borrowed books database. Try typing "I would like to borrow Harry Potter"
- list: Returns a list of your borrowed books. Try typing "I want to see my borrowed list"
- return {book title}: returns a book from your borrowed books list. If a book is not in the list, it will inform you of that. Try typing "I would like to return Harry Potter "
- reco {genre}:  Recommends a book from a particular genre. Try typing "recommend scifi"
- search for {book name}: Returns author, cover and description for a particular book. Try typing "search for Harry Potter"
- make review: Makes a review of a particular book by asking the user a series of questions. Try typing "make review"
- get all reviews: returns a list of all the reviews made by users of the bot. Try typing "get all reviews"