projectile motion simulator using pygame
Simulates the flight path of an object given initial velocity and launch angle. This project was made to help with physics homework. Right and left arrows control launch angle, and up and down arrows control velocity. Still a bit laggy because the screen has to be refreshed on every key press, but I'm working on fixing that.